Deny the user to change page geometry
In this stage of development the page dimensions are hardcoded
(to A4-landscape, equaling 297x210 mm). Thus, it does not make sense
to allow the user to change these dimensions yet.
This commit makes it clear to the user that changing the page geometry
is not yet supported. Warnings are issued if the page geometry is being
changed, and the values are automatically corrected to the hardcoded
ones. This is to minimize user confusion.